backend/requirements.txt
flask
firebase-admin
flask-cors
python-dotenv
requests
Flask-Limiter
bleach
Flask-Talisman
sentry-sdk[flask]
gunicorn
python-slugify
